@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { FormGroup, FormControl, Validators } from '@angular/forms';
 
 @Component({
   selector: 'app-register',
@@ -6,10 +7,20 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./register.component.css']
 })
 export class RegisterComponent implements OnInit {
+  forma: FormGroup;
 
-  constructor() { }
-
-  ngOnInit() {
+  constructor() {
+    this.forma = new FormGroup({
+      nombre: new FormControl('', [Validators.required, Validators.minLength(3)]),
+      apellido: new FormControl('', Validators.required),
+      correo: new FormControl('', [Validators.required, Validators.pattern('[a-z0-9._%+-]+@[a-z0-9.-]+.[a-z]{2,3}$')])
+    });
   }
 
+  ngOnInit() {}
+
+  guardar() {
+    console.log(this.forma.value);
+    console.log(this.forma);
+  }
 }
