@@ -1,5 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import { AuthService } from '../../../services/auth.service';
+import { log } from 'util';
 
 @Component({
   selector: 'app-navbar',
@@ -7,9 +9,29 @@ import { Router } from '@angular/router';
   styleUrls: ['./navbar.component.scss']
 })
 export class NavbarComponent implements OnInit {
-  constructor(private router: Router) {}
+  public logueado: string;
+  constructor(private router: Router, public auth: AuthService) {
+    auth.handleAuthentication();
+  }
 
-  ngOnInit() {}
+  ngOnInit() {
+    // console.log('Hola');
+    this.logueado = localStorage.getItem('isLoggedIn');
+    console.log('Logueado = ' + localStorage.getItem('isLoggedIn'));
+  }
+
+  login() {
+    this.auth.login();
+  }
+
+  salir() {
+    this.auth.logout();
+  }
+
+  account() {
+    this.router.navigate(['/account']);
+  }
+
   buscarProducto(textoSearch: string) {
     console.log(textoSearch);
     this.router.navigate(['/resultado', textoSearch]);
